% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_PIVQMLE.R
\name{A_PIVQMLE}
\alias{A_PIVQMLE}
\title{Adaptive quasi-maximum likelihood estimation}
\usage{
A_PIVQMLE(series, order = c(1, 1))
}
\arguments{
\item{series}{The original time-series that need to be fitted as GARCH model}

\item{order}{The GARCH model orders. Includes GARCH terms and ARCH terms.}
}
\value{
Estimated parameters.
}
\description{
This estimation method works for GARCH(p,q) model with Pearson's type IV innovation.
}

