% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_tQMLE.R
\name{A_tQMLE}
\alias{A_tQMLE}
\title{Adaptive Quasi Maximum Likelihood Estimation}
\usage{
A_tQMLE(series, order = c(1, 1))
}
\arguments{
\item{series}{The original time-series that need to be fitted as GARCH model}

\item{order}{The GARCH model orders. Includes GARCH terms and ARCH terms.}
}
\value{
The estimation of parameters.
}
\description{
This is a function which can be used to estimate parameters of all kinds of models.
The critical idea of this methods is to estimate parameters according to data
themselves with unspecified assumption of error's distribution.
}
\examples{
#test
}

