% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.residue.R
\name{com.residue}
\alias{com.residue}
\title{Internal function to calculate estimation residuals}
\usage{
com.residue(alpha, beta, series)
}
\arguments{
\item{alpha}{A vector includes intercept and all parameters of ARCH terms.}

\item{beta}{A vector includes all parameters of GARCH terms.}

\item{series}{Original time series.}
}
\value{
Return a list contains estimators, residues and conditional sigma square.
}
\description{
A internal function used to calculate the residue of GARCH model so that the futher adaptive procedure can be move forward.
}

