% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARCH_Gene.R
\name{GARCH_Gene}
\alias{GARCH_Gene}
\title{GARCH_Gene}
\usage{
GARCH_Gene(alpha, beta, n = 100, rnd = c("rt", "rnorm", "rpearsonIV"), df.t,
  params.PIV, ntrans = 100)
}
\arguments{
\item{alpha}{vector include the intercept and the parameters of autoregression part, which is ARCH terms (q).}

\item{beta}{vector include the parameters of conditional variance, which is GARCH terms with order p.}

\item{n}{lenght of time series}

\item{rnd}{distribution of error(innovation) of GARCH model. Innovation distribution options includes 
student's t, normal distributions, generalized Gaussian and Pearson Type IV distributions}

\item{df.t}{degree of freedom of student's t distribution. df.t>2 continuous parameter}

\item{params.PIV}{vector/list of length 4 containing parameters m, nu, location, scale for Pearson 
type IV distribution (in this order!).}

\item{ntrans}{burn-in size, i.e. number of initial simulated data to be discarded}
}
\value{
A list includes time-series "x" and conditional sigma square "sig.sq".
}
\description{
Function to generate GARCH time series with normal, student's t and person type IV distribution 
innovations
}
\details{
This function is used to generate GARCH time series with t-students distribution innovations. 
This kind of innovations will better perform the heavy-tailed characteristics of financial data.
}
\examples{
# Test for GARCH(1,1) with t innovation #
xx <- GARCH_Gene(alpha = c(0.1, 0.4), beta = 0.4, n = 2000, rnd = "rt", df.t = 4)
y <- xx$x
plot(y, type = "l")
paraPIV <- list(m=2, nu=-2, location=0, scale=1)
x.PIV <- GARCH_Gene(alpha = c(0.1, 0.2), beta = 0.5, n = 1000, rnd = "rpearsonIV", 
                    params.PIV = paraPIV)
\dontrun{hist(x.PIV$x)}
\dontrun{plot(x.PIV$x, type = 'l')}
}
\author{
Xiaorui(Jeremy) Zhu
}

