% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.lin.R
\name{gen.lin}
\alias{gen.lin}
\title{Single linear model y = a + b*x.}
\usage{
gen.lin(intercept, slope, noise.variance, x, dis.error = c("rt", "rnorm"),
  dft)
}
\arguments{
\item{intercept}{The intercept of single linear model.}

\item{slope}{The parameter b, which is the slope of y = a + b*x.}

\item{noise.variance}{The variance of error series.}

\item{x}{The depedent variable.}

\item{dis.error}{Distribution of error, student's t or normal.}

\item{dft}{The degree of freedom if error was specified as student's t.}
}
\value{
A data frame contains x, y and error series.
}
\description{
Generating function for single linear model y = a + b*x.
}

