% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogL_GARCH_Norm.R
\name{LogL_GARCH_Norm}
\alias{LogL_GARCH_Norm}
\title{Log-likelihood function fo GARCH(p,q)}
\usage{
LogL_GARCH_Norm(series, p, q)
}
\arguments{
\item{series}{The time series that need to be fitted.}

\item{p}{The order of GARCH terms \eqn{\sigma^2}}

\item{q}{The order of ARCH terms \eqn{\epsilon^2}}
}
\value{
Log likelihood function
}
\description{
The Log-likelihood function of GARCH(p,q) model.
}

