% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogL_GARCH_t.R
\name{LogL_GARCH_t}
\alias{LogL_GARCH_t}
\title{LogL_GARCH_t}
\usage{
LogL_GARCH_t(series, p, q, df)
}
\arguments{
\item{series}{The time series that need to be fitted}

\item{p}{The order of GARCH terms \eqn{\sigma^2}}

\item{q}{The order of ARCH terms \eqn{\epsilon^2}}

\item{df}{The specified degree of freedome of student's t innovation.}
}
\value{
Log likelihood funtion.
}
\description{
The Log likelihood function of GARCH model. Here is the GARCH(p,q) with student's t error
}

