% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YITAtQMLLE.R
\name{YITAtQMLE}
\alias{YITAtQMLE}
\title{Estimation of scale parameter \eqn{\eta_f} using tQMLE}
\usage{
YITAtQMLE(e, dfest)
}
\arguments{
\item{e}{The innovation caculated by estimate parameters}

\item{dfest}{The parameter of Quais likelihood. Continuous parameter dfest>2 so that finite variance}
}
\value{
The estimator of eta
}
\description{
This function is used to estimate scale parameter \eqn{\eta_f} from the model residues under specified assumption of distribution of error. So the input includes degree of freedom and model residues.
}
\examples{
# need revisions
}

