% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogL_GARCH_PIV.R
\name{LogL_GARCH_PIV}
\alias{LogL_GARCH_PIV}
\title{log-likelihood function of GARCH(p,q) with Pearson's Type IV (PIV) distribution}
\usage{
LogL_GARCH_PIV(series, p, q, df)
}
\arguments{
\item{series}{The time series that need to be fitted.}

\item{p}{The order of GARCH terms \eqn{\sigma^2}}

\item{q}{The order of ARCH terms \eqn{\epsilon^2}}

\item{df}{The specified degree of freedome of student's t innovation.}
}
\value{
Log-Likelihood function of GARCH with t error
}
\description{
The Log-likelihood function of GARCH model with Pearson's Type IV (PIV) distributed error
}

